pyproject.toml
[project]
name = "evolutionapi-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.58.1",
    "pycryptodome>=3.21.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "whisper>=1.1.10",
]
